+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Add support to interactive messages. [#165](https://github.com/DevAlexandreCR/gorda-admin/pull/165)
 
+- Add go back button to drivers. [#167](https://github.com/DevAlexandreCR/gorda-admin/pull/167)
+
 # Release Notes for 1.5.x
 
 ## [1.5.5(2025-05-13)](https://github.com/DevAlexandreCR/admin-driver/compare/1.5.5...1.5.4)